+++ src/helpers/CkeditorConfig.php
@@ -8,6 +8,7 @@
 namespace craft\ckeditor\helpers;
 
 use Craft;
+use craft\ckeditor\models\EntryType as CkeEntryType;
 use Illuminate\Support\Collection;
 
 /**
@@ -411,5 +412,20 @@ final class CkeditorConfig
                 ],
             ],
         ];
+    }
+
+    /**
+     * Get ckeditor's Entry Type based on provided config.
+     *
+     * @param array $config
+     * @return CkeEntryType
+     * @since 5.0.0
+     */
+    public static function getCkeEntryType(array $config): CkeEntryType
+    {
+        $craftEntryType = Craft::$app->getEntries()->getEntryType($config);
+        $properties = get_object_vars($craftEntryType) + $config;
+
+        return new CkeEntryType($properties);
     }
 }