+++ src/web/assets/ckeconfig/src/ckeconfig.js
@@ -6,3 +6,4 @@
 
 export {default as ToolbarBuilder} from './ToolbarBuilder.js';
 export {default as ConfigOptions} from './ConfigOptions.js';
+export {default as CkeEntryTypeSelectInput} from './CkeEntryTypeSelectInput.js';